Renaming FieldModelPartP{CS} into FieldModelTextureRefP{CS}.
