+++ Makou_Reactor.pro
@@ -87,7 +87,6 @@ HEADERS += \
     core/field/GrpScript.h \
     core/field/FieldPS.h \
     core/field/FieldPC.h \
-    core/field/FieldModelPartPS.h \
     core/field/FieldModelPart.h \
     core/field/FieldModelLoaderPS.h \
     core/field/FieldModelLoaderPC.h \
@@ -137,7 +136,9 @@ HEADERS += \
     core/field/AFile.h \
     core/field/PFile.h \
     core/field/RsdFile.h \
-    core/field/FieldModelPartPC.h
+    core/field/FieldModelTextureRef.h \
+    core/field/FieldModelTextureRefPC.h \
+    core/field/FieldModelTextureRefPS.h
 
 SOURCES += \
     Window.cpp \
@@ -218,7 +219,6 @@ SOURCES += \
     core/field/GrpScript.cpp \
     core/field/FieldPS.cpp \
     core/field/FieldPC.cpp \
-    core/field/FieldModelPartPS.cpp \
     core/field/FieldModelPart.cpp \
     core/field/FieldModelLoaderPS.cpp \
     core/field/FieldModelLoaderPC.cpp \
@@ -268,7 +268,9 @@ SOURCES += \
     core/field/AFile.cpp \
     core/field/PFile.cpp \
     core/field/RsdFile.cpp \
-    core/field/FieldModelPartPC.cpp
+    core/field/FieldModelTextureRef.cpp \
+    core/field/FieldModelTextureRefPC.cpp \
+    core/field/FieldModelTextureRefPS.cpp
 
 TRANSLATIONS += Makou_Reactor_en.ts \
     Makou_Reactor_ja.ts
