+++ core/field/FieldModelTextureRefPC.cpp
@@ -15,4 +15,4 @@
  ** You should have received a copy of the GNU General Public License
  ** along with this program.  If not, see <http://www.gnu.org/licenses/>.
  ****************************************************************************/
-#include "FieldModelPartPC.h"
+#include "FieldModelTextureRefPC.h"
