+++ core/field/PFile.h
@@ -20,7 +20,7 @@
 
 #include <QtCore>
 #include "../IO.h"
-#include "FieldModelPartPC.h"
+#include "FieldModelPart.h"
 #include "IdFile.h"
 
 #define MODEL_SCALE_PC				132.0f // 1.0f
